Renamed NameSpace and changed to a single DLL that works regardless of 32 or 64 bit OS.
