+++ Computer.Info.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.2015
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WiensWare.Computer.Info", "WiensWare.Computer.Info.csproj", "{495181F3-F508-44F8-86E6-14F67998303E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test", "Test\Test.csproj", "{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{495181F3-F508-44F8-86E6-14F67998303E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{495181F3-F508-44F8-86E6-14F67998303E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{495181F3-F508-44F8-86E6-14F67998303E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{495181F3-F508-44F8-86E6-14F67998303E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {FA43A8D9-517B-473C-92E8-BB72966C69F8}
+	EndGlobalSection
+EndGlobal
