--- ComputerInfo.sln
@@ -1,41 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27428.2015
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ComputerInfo", "ComputerInfo.csproj", "{495181F3-F508-44F8-86E6-14F67998303E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test", "Test\Test.csproj", "{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{495181F3-F508-44F8-86E6-14F67998303E}.Debug|x64.ActiveCfg = Debug|x64
-		{495181F3-F508-44F8-86E6-14F67998303E}.Debug|x64.Build.0 = Debug|x64
-		{495181F3-F508-44F8-86E6-14F67998303E}.Debug|x86.ActiveCfg = Debug|x86
-		{495181F3-F508-44F8-86E6-14F67998303E}.Debug|x86.Build.0 = Debug|x86
-		{495181F3-F508-44F8-86E6-14F67998303E}.Release|x64.ActiveCfg = Release|x64
-		{495181F3-F508-44F8-86E6-14F67998303E}.Release|x64.Build.0 = Release|x64
-		{495181F3-F508-44F8-86E6-14F67998303E}.Release|x86.ActiveCfg = Release|x86
-		{495181F3-F508-44F8-86E6-14F67998303E}.Release|x86.Build.0 = Release|x86
-		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Debug|x64.ActiveCfg = Debug|x64
-		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Debug|x64.Build.0 = Debug|x64
-		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Debug|x86.ActiveCfg = Debug|x86
-		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Debug|x86.Build.0 = Debug|x86
-		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Release|x64.ActiveCfg = Release|x64
-		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Release|x64.Build.0 = Release|x64
-		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Release|x86.ActiveCfg = Release|x86
-		{7C6C8C4A-3882-4795-B6B4-79032FCE6AC7}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FA43A8D9-517B-473C-92E8-BB72966C69F8}
-	EndGlobalSection
-EndGlobal
